--- starter/src/test/java/de/fraunhofer/iosb/ilt/faaast/service/starter/AppPropertyOverrideTest.java
@@ -1,115 +1,0 @@
-/*
- * Copyright (c) 2021 Fraunhofer IOSB, eine rechtlich nicht selbstaendige
- * Einrichtung der Fraunhofer-Gesellschaft zur Foerderung der angewandten
- * Forschung e.V.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package de.fraunhofer.iosb.ilt.faaast.service.starter;
-
-import de.fraunhofer.iosb.ilt.faaast.service.config.ServiceConfig;
-import de.fraunhofer.iosb.ilt.faaast.service.starter.util.ParameterConstants;
-import de.fraunhofer.iosb.ilt.faaast.service.starter.util.ServiceConfigHelper;
-import java.io.File;
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-import org.junit.*;
-import picocli.CommandLine;
-
-
-public class AppPropertyOverrideTest extends AbstractAppTest {
-    private static ServiceConfig config;
-    private static final String CONFIG = "src/test/resources/config-dummy-messagebus.json";
-
-    Map<String, String> envProperties;
-
-    @BeforeClass
-    public static void init() throws IOException {
-        config = ServiceConfigHelper.load(new File(CONFIG));
-    }
-
-
-    @Before
-    public void initEnvProperties() {
-        envProperties = new HashMap<>();
-        envProperties.put(App.ENV_CONFIG_FILE_PATH, CONFIG);
-    }
-
-
-    @Test
-    public void testSeparatorReplacement() throws Exception {
-        Map<String, String> expected = new HashMap<>();
-        expected.put(ParameterConstants.MESSAGEBUS_NO_UNDERSCORE_AFTER, "1");
-        expected.put(ParameterConstants.MESSAGEBUS_UNDERSCORE_AFTER, "1");
-
-        Map<String, String> envProperties = new HashMap<>();
-        envProperties.put(ParameterConstants.MESSAGEBUS_NO_UNDERSCORE_BEFORE, "1");
-        envProperties.put(ParameterConstants.MESSAGEBUS_UNDERSCORE_BEFORE, "1");
-
-        Map<String, String> actual = withEnv(envProperties).execute(() -> {
-            new CommandLine(application).execute();
-            return application.getConfigOverrides(config);
-        });
-        Assert.assertEquals(expected, actual);
-    }
-
-
-    @Test
-    public void testNestedSeparatorReplacement() throws Exception {
-        Map<String, String> expected = new HashMap<>();
-        expected.put(ParameterConstants.MESSAGEBUS_NESTED_NO_UNDERSCORE_AFTER, "1");
-        expected.put(ParameterConstants.MESSAGEBUS_NESTED_UNDERSCORE_AFTER, "1");
-
-        Map<String, String> envProperties = new HashMap<>();
-        envProperties.put(ParameterConstants.MESSAGEBUS_NESTED_NO_UNDERSCORE_BEFORE, "1");
-        envProperties.put(ParameterConstants.MESSAGEBUS_NESTED_UNDERSCORE_BEFORE, "1");
-
-        Map<String, String> actual = withEnv(envProperties).execute(() -> {
-            new CommandLine(application).execute();
-            return application.getConfigOverrides(config);
-        });
-        Assert.assertEquals(expected, actual);
-    }
-
-
-    @Test
-    public void testPraefixSeparatorReplacement() throws Exception {
-        Map<String, String> expected = new HashMap<>();
-        expected.put(ParameterConstants.MESSAGEBUS_PRAEFIX_AFTER, "1");
-
-        Map<String, String> envProperties = new HashMap<>();
-        envProperties.put(ParameterConstants.MESSAGEBUS_PRAEFIX_BEFORE, "1");
-
-        Map<String, String> actual = withEnv(envProperties).execute(() -> {
-            new CommandLine(application).execute();
-            return application.getConfigOverrides(config);
-        });
-        Assert.assertEquals(expected, actual);
-    }
-
-
-    @Test
-    public void testAmbiguitySeparatorReplacement() throws Exception {
-        Map<String, String> envProperties = new HashMap<>();
-        envProperties.put(ParameterConstants.MESSAGEBUS_AMBIGUITY_BEFORE, "1");
-
-        try {
-            withEnv(envProperties).execute(() -> {
-                new CommandLine(application).execute();
-                return application.getConfigOverrides(config);
-            });
-            Assert.fail();
-        }
-        catch (InitializationException e) {
-            Assert.assertTrue(true);
-        }
-    }
-}
