+++ starter/src/test/java/de/fraunhofer/iosb/ilt/faaast/service/starter/AppTest.java
@@ -14,29 +14,50 @@
  */
 package de.fraunhofer.iosb.ilt.faaast.service.starter;
 
+import de.fraunhofer.iosb.ilt.faaast.service.config.ServiceConfig;
 import de.fraunhofer.iosb.ilt.faaast.service.endpoint.http.HttpEndpoint;
+import de.fraunhofer.iosb.ilt.faaast.service.starter.fixtures.DummyMessageBusConfig;
 import de.fraunhofer.iosb.ilt.faaast.service.starter.util.ParameterConstants;
+import de.fraunhofer.iosb.ilt.faaast.service.util.Ensure;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.PrintWriter;
+import java.io.StringWriter;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+import java.util.stream.Stream;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.Test;
 import picocli.CommandLine;
+import uk.org.webcompere.systemstubs.environment.EnvironmentVariables;
 
 
-public class AppTest extends AbstractAppTest {
+public class AppTest {
 
+    protected App application;
+    protected CommandLine cmd;
     private static final String MODEL_RESOURCE_PATH = "/AASMinimal.json"; // Path of model resource from core dependency
     private static final String CONFIG = "src/test/resources/config-minimal.json";
     private Path modelPath;
+    private static ServiceConfig dummyMessageBusConfig;
+
+    @BeforeClass
+    public static void init() {
+        DummyMessageBusConfig messageBusConfig = new DummyMessageBusConfig();
+        dummyMessageBusConfig = ServiceConfig.builder()
+                .messageBus(messageBusConfig)
+                .build();
+    }
+
 
     @Before
     public void prepareResources() throws IOException {
@@ -44,6 +65,49 @@ public class AppTest extends AbstractAppTest {
         InputStream modelResourceAsStream = AppTest.class.getResourceAsStream(MODEL_RESOURCE_PATH);
 
         Files.copy(modelResourceAsStream, modelPath);
+    }
+
+
+    @Before
+    public void initCmd() throws IOException {
+        application = new App();
+        application.dryRun = true;
+        cmd = new CommandLine(application)
+                .setCaseInsensitiveEnumValuesAllowed(true);
+        cmd.setOut(new PrintWriter(new StringWriter()));
+    }
+
+
+    protected EnvironmentVariables withEnv(Map<String, String> variables) {
+        return withEnv(variables.entrySet().stream()
+                .map(x -> new String[] {
+                        x.getKey(),
+                        x.getValue()
+                })
+                .flatMap(x -> Stream.of(x))
+                .toArray(String[]::new));
+    }
+
+
+    protected EnvironmentVariables withEnv(String... variables) {
+        Ensure.requireNonNull(variables, "variables must be non-null");
+        Ensure.require(variables.length >= 2, "variables must contain at least one element");
+        Ensure.require(variables.length % 2 == 0, "variables must contain an even number of elements");
+
+        EnvironmentVariables result = null;
+        for (int i = 0; i < variables.length; i += 2) {
+            String key = variables[i];
+            if (!Objects.equals(App.ENV_CONFIG_FILE_PATH, key) && !Objects.equals(App.ENV_MODEL_FILE_PATH, key)) {
+                key = key.startsWith(App.ENV_CONFIG_EXTENSION_PREFIX)
+                        ? key
+                        : String.format("%s%s", App.ENV_CONFIG_EXTENSION_PREFIX, key);
+            }
+            String value = variables[i + 1];
+            result = result == null
+                    ? new EnvironmentVariables(key, value)
+                    : result.and(key, value);
+        }
+        return result;
     }
 
 
@@ -71,6 +135,68 @@ public class AppTest extends AbstractAppTest {
             return application.getConfigOverrides();
         });
         Assert.assertEquals(expected, actual);
+    }
+
+
+    @Test
+    public void testSeparatorReplacement() throws Exception {
+        Map<String, String> expected = new HashMap<>();
+        expected.put(ParameterConstants.MESSAGEBUS_NO_UNDERSCORE_AFTER, "1");
+        expected.put(ParameterConstants.MESSAGEBUS_UNDERSCORE_AFTER, "1");
+
+        Map<String, String> envProperties = new HashMap<>();
+        envProperties.put(ParameterConstants.MESSAGEBUS_NO_UNDERSCORE_BEFORE, "1");
+        envProperties.put(ParameterConstants.MESSAGEBUS_UNDERSCORE_BEFORE, "1");
+
+        Map<String, String> actual = withEnv(envProperties).execute(() -> {
+            return application.getConfigOverrides(dummyMessageBusConfig);
+        });
+        Assert.assertEquals(expected, actual);
+    }
+
+
+    @Test
+    public void testNestedSeparatorReplacement() throws Exception {
+        Map<String, String> expected = new HashMap<>();
+        expected.put(ParameterConstants.MESSAGEBUS_NESTED_NO_UNDERSCORE_AFTER, "1");
+        expected.put(ParameterConstants.MESSAGEBUS_NESTED_UNDERSCORE_AFTER, "1");
+
+        Map<String, String> envProperties = new HashMap<>();
+        envProperties.put(ParameterConstants.MESSAGEBUS_NESTED_NO_UNDERSCORE_BEFORE, "1");
+        envProperties.put(ParameterConstants.MESSAGEBUS_NESTED_UNDERSCORE_BEFORE, "1");
+
+        Map<String, String> actual = withEnv(envProperties).execute(() -> {
+            return application.getConfigOverrides(dummyMessageBusConfig);
+        });
+        Assert.assertEquals(expected, actual);
+    }
+
+
+    @Test
+    public void testPraefixSeparatorReplacement() throws Exception {
+        Map<String, String> expected = new HashMap<>();
+        expected.put(ParameterConstants.MESSAGEBUS_PRAEFIX_AFTER, "1");
+
+        Map<String, String> envProperties = new HashMap<>();
+        envProperties.put(ParameterConstants.MESSAGEBUS_PRAEFIX_BEFORE, "1");
+
+        Map<String, String> actual = withEnv(envProperties).execute(() -> {
+            return application.getConfigOverrides(dummyMessageBusConfig);
+        });
+        Assert.assertEquals(expected, actual);
+    }
+
+
+    @Test
+    public void testAmbiguitySeparatorReplacement() throws Exception {
+        Map<String, String> envProperties = new HashMap<>();
+        envProperties.put(ParameterConstants.MESSAGEBUS_AMBIGUITY_BEFORE, "1");
+
+        Assert.assertThrows(InitializationException.class, () -> {
+            withEnv(envProperties).execute(() -> {
+                application.getConfigOverrides(dummyMessageBusConfig);
+            });
+        });
     }
